+++ packages/react/src/index.ts
@@ -15,6 +15,6 @@ export { BasicTab, BorderTab } from './components/Tab';
 export { ColumnTable, RowTable, BasicTable } from './components/Table';
 export * from './components/Text';
 export * from './components/Icon';
-export { default as Slider } from './components/Slider';
+export * from './components/Slider';
 export { default as cosStyle, CosStyle } from './styles/GlobalStyle';
 export type { Theme, Palette } from './styles/theme';
